requirements.txt
requests
beautifulsoup4
flask
python-whois
#nmap/zmap?
